Create Release01
First Project: The Map Generator
This first release of the game is just the foundation for the exciting modifications to come.  It is basically a map generator, which can be played as a freestanding exploration game.

This code uses on-the-fly map generation (dynamic map generation) to allow unlimited exploration.

  [see screenshot at http://www.snowflakefuture.com/2016/10/first-project-map-generator.html ]

The sheets/tabs of your excel file must be named as in the above screenshot.  You must also add buttons to call the macros "ResumeGame" and "ShowStartButton".  And you must create user forms as shown at the bottom of the code section.

Finally, you should include text on rows 6 and 7 of the sheet/tab labeled "Map" as in the following screenshot:

  [see screenshot at http://www.snowflakefuture.com/2016/10/first-project-map-generator.html ]

To play single player mode, go into the sheet/tab labeled "Map" and delete "human" from row 7.  To add a third player, create two additional sheets/tabs ("Player3" and "Player3 Hex View"), and add "human" and a color index to row 8.